pyproject.toml
[project]
name = "news-discuss"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "langgraph>=0.2.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "chromadb>=0.4.0",
    "httpx>=0.25.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "openai>=1.0.0",
    "numpy>=1.24.0",
    "sentence-transformers>=2.2.0",
    "uuid>=1.30",
]
